refactor: lots of cleanup. one failing test remaining
