+++ std/compress/lzss_v1/writeCs/main.go
@@ -1,12 +1,18 @@
 package main
 
 import (
+	"bytes"
+	"compress/gzip"
 	"fmt"
 	"github.com/consensys/gnark-crypto/ecc"
+	"github.com/consensys/gnark/constraint"
 	"github.com/consensys/gnark/frontend"
 	"github.com/consensys/gnark/frontend/cs/scs"
 	"github.com/consensys/gnark/std/compress/lzss_v1"
+	"os"
 )
+
+// executable to generate the constraints for the circuit and store it
 
 func check(e error) {
 	if e != nil {
@@ -27,11 +33,35 @@ func main() {
 		},
 	}
 
-	//p := profile.Start()
 	cs, err := frontend.Compile(ecc.BN254.ScalarField(), scs.NewBuilder, &c)
 	check(err)
 
-	//p.Stop()
 	fmt.Println(cs.GetNbConstraints(), "constraints")
-	check(lzss_v1.GzCompressCs("600kb.gz", cs))
+	check(gzCompressCs("600kb.gz", cs))
+}
+
+func gzCompressCs(outFileName string, cs constraint.ConstraintSystem) error {
+	var raw bytes.Buffer
+	_, err := cs.WriteTo(&raw)
+	if err != nil {
+		return err
+	}
+	compressed, err := gzCompress(raw.Bytes())
+	if err != nil {
+		return err
+	}
+	return os.WriteFile(outFileName, compressed, 0644)
+}
+
+func gzCompress(in []byte) ([]byte, error) {
+	var out bytes.Buffer
+	w := gzip.NewWriter(&out)
+	_, err := w.Write(in)
+	if err != nil {
+		return nil, err
+	}
+	if err = w.Close(); err != nil {
+		return nil, err
+	}
+	return out.Bytes(), nil
 }
