+++ std/compress/test_cases/3c2943/data.zfct
