+++ std/compress/test_cases/705b24/candidates.txt
