+++ std/compress/test_cases/705b24/data.zfct
