+++ std/compress/test_cases/777003/data.zct
