+++ std/compress/test_cases/bug/data.bin
