+++ std/compress/test_cases/bug/data.lzssv1
