+++ std/compress/test_cases/large/data.lzssv1
